move test cases for claim_parser to their own file
